Rename penaltymodel.cache to penaltymodel.database
