+++ tests/test_interface.py
@@ -20,7 +20,7 @@ import dimod
 import networkx as nx
 
 from penaltymodel import get_penalty_model
-from penaltymodel.cache import isolated_cache
+from penaltymodel.database import isolated_cache
 
 
 class TestGetPenaltyModel(unittest.TestCase):
